requirements.txt
anthropic>=0.40.0
pymongo[srv]>=4.7.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
PyMuPDF>=1.24.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
python-multipart>=0.0.9
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
